add favicon and fix navbar
