--- app/javascript/plugins/navbarShow.js
@@ -1,32 +1,0 @@
-const navbarShow = () => {
-  const map = document.getElementById("map");
-  if (map) {
-    const navbar = document.getElementById('navbar');
-    const toggle = document.getElementById('toggle');
-    const toggleIcon = document.getElementById('menu-icon');
-    document.addEventListener("scroll", (e) => {
-      if (navbar.offsetHeight < window.scrollY) {
-        navbar.classList.remove('naked-navbar');
-        navbar.style.position = 'fixed';
-        navbar.style.width = '100%';
-      } else {
-        navbar.classList.add('naked-navbar');
-      }
-    });
-    toggle.addEventListener('click', (e) => {
-      if (toggle.getAttribute('aria-expanded') === 'false') {
-        navbar.classList.remove('naked-navbar');
-        navbar.style.position = 'fixed';
-        navbar.style.width = '100%';
-        toggleIcon.classList.add('fa-times');
-        toggleIcon.classList.remove('fa-bars');
-      } else {
-        navbar.classList.add('naked-navbar');
-        toggleIcon.classList.add('fa-bars');
-        toggleIcon.classList.remove('fa-times');
-      }
-    });
-  }
-};
-
-export { navbarShow };
